+++ docs/CHANGELOG.md
@@ -196,7 +196,7 @@ General:
    - Added support for particube `pcubes` and `3zh` support
    - Added support for Photoshop `ase` palette
    - Added support for Paintshop Pro (JASC) `pal` palette
-   - Extended support for the Gimp `gpl` palette format to support the AseSprite alpha channel extension
+   - Extended support for the Gimp `gpl` palette format to support the Aseprite alpha channel extension
    - Added support to collect anonymous metrics (disabled by default)
    - Added support to align all models in a scene - useful for rendering all models for an overview
    - Extended lua bindings